chore: move IGameState.ts into types dir
