+++ app/frontend/src/types/IGameState.ts
